requirements.txt
python-binance
numpy
TA-Lib
websocket-client